refactor: remove spec files in the components
