+++ src/pages/CoinsPage/CoinsPage.spec.tsx
